MAT-7401 refactored code and added test cases
